Standardize "common helpers" for command trees
Rather than only having global level organization for shared parsing
in `parsing.shared_options`, standardize on `_common.py` as the name
for a subpackage-local module for helpers. Several global options like
task_id can be moved to local helpers to reduce the bloat in the
global helper module.

This also establishes a semi-standard way of organizing the helpers
relevant for a family of commands.
