+++ src/globus_cli/commands/endpoint/role/_common.py
@@ -1,0 +1,5 @@
+import click
+
+
+def role_id_arg(f):
+    return click.argument("role_id")(f)
