+++ src/globus_cli/commands/endpoint/role/delete.py
@@ -1,7 +1,9 @@
 from globus_cli.login_manager import requires_login
-from globus_cli.parsing import command, endpoint_id_arg, role_id_arg
+from globus_cli.parsing import command, endpoint_id_arg
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
 from globus_cli.termio import FORMAT_TEXT_RAW, formatted_print
+
+from ._common import role_id_arg
 
 
 @command(
