+++ src/globus_cli/commands/endpoint/server/add.py
@@ -1,11 +1,13 @@
 from globus_cli.login_manager import requires_login
-from globus_cli.parsing import command, endpoint_id_arg, server_add_and_update_opts
+from globus_cli.parsing import command, endpoint_id_arg
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     assemble_generic_doc,
     get_client,
 )
 from globus_cli.termio import FORMAT_TEXT_RAW, formatted_print
+
+from ._common import server_add_and_update_opts
 
 
 @command(
