+++ src/globus_cli/commands/task/event_list.py
@@ -3,7 +3,7 @@ import json
 import click
 
 from globus_cli.login_manager import requires_login
-from globus_cli.parsing import command, task_id_arg
+from globus_cli.parsing import command
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     get_client,
@@ -11,6 +11,8 @@ from globus_cli.services.transfer import (
 )
 from globus_cli.termio import formatted_print
 from globus_cli.utils import PagingWrapper
+
+from ._common import task_id_arg
 
 
 @command(
