+++ src/globus_cli/parsing/__init__.py
@@ -1,27 +1,22 @@
 from globus_cli.parsing.commands import command, group, main_group
 from globus_cli.parsing.explicit_null import EXPLICIT_NULL
-from globus_cli.parsing.mutex_group import mutex_option_group
+from globus_cli.parsing.mutex_group import MutexInfo, mutex_option_group
 from globus_cli.parsing.one_use_option import one_use_option
 from globus_cli.parsing.process_stdin import shlex_process_stdin
 from globus_cli.parsing.shared_options import (
     delete_and_rm_options,
-    endpoint_create_and_update_params,
     endpoint_id_arg,
     no_local_server_option,
-    role_id_arg,
     security_principal_opts,
-    server_add_and_update_opts,
-    server_id_arg,
     synchronous_task_wait_options,
-    task_id_arg,
     task_submission_options,
-    validate_endpoint_create_and_update_params,
 )
 
 from .param_types import (
     ENDPOINT_PLUS_OPTPATH,
     ENDPOINT_PLUS_REQPATH,
     IdentityType,
+    LocationType,
     TaskPath,
 )
 
@@ -36,21 +31,17 @@ __all__ = [
     "ENDPOINT_PLUS_REQPATH",
     "TaskPath",
     "one_use_option",
+    "MutexInfo",
     "mutex_option_group",
     "IdentityType",
+    "LocationType",
     # other helpers
     "EXPLICIT_NULL",
     # Transfer options
     "endpoint_id_arg",
-    "task_id_arg",
     "task_submission_options",
     "delete_and_rm_options",
     "synchronous_task_wait_options",
-    "endpoint_create_and_update_params",
-    "validate_endpoint_create_and_update_params",
-    "role_id_arg",
-    "server_id_arg",
-    "server_add_and_update_opts",
     "security_principal_opts",
     "no_local_server_option",
     "shlex_process_stdin",
